pyproject.toml
[tool.poetry]
name = "twallet"
version = "0.1.0"
description = "Telegram Ton Wallet"
authors = ["yoyo <hvuhsg5@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tonsdk = "^1.0.13"
tvm-valuetypes = "^0.0.12"
aiohttp = "^3.8.5"
python-telegram-bot = "^20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
